stablediff_app/requirements.txt
torch >= 2.0.0
transformers >= 4.36.0
diffusers >= 0.27.0
streamlit == 1.31.1
Pillow
accelerate
safetensors
scikit-learn
torchvision